Invert logic for P Chan mosfets, 0v will activate them. Previously handled with a BJT on board. Now using gate drivers without an inverted output. Need to make this clearer. False to arm is confusing.

+++ src/framework.cpp
@@ -114,7 +114,7 @@ int Framework::DigitalRead(int Pin)
 
 //-------------------------------------------------------------------------------
 //-------------------------------------------------------------------------------
-void Framework::AnalogWrite(int Pin, int8_t PwmValue)
+void Framework::AnalogWrite(int Pin, int PwmValue)
 {
 	//Framework::Message("Pin=%d - Value=%d", Pin, (int)bState);
 	analogWrite(Pin, PwmValue);
